+++ lib/td_dd_web/controllers/graph_controller.ex
@@ -35,6 +35,25 @@ defmodule TdDdWeb.GraphController do
     end
   end
 
+  def csv(conn, %{"id" => id}) do
+    with %Graph{data: data} <- Graphs.get!(id) do
+      type =
+        data
+        |> Map.get("opts")
+        |> Map.get("type")
+
+      attrs =
+        Map.new()
+        |> Map.put("type", type)
+        |> Map.merge(Map.take(data, ["ids", "excludes"]))
+
+      conn
+      |> put_resp_content_type("text/csv", "utf-8")
+      |> put_resp_header("content-disposition", "attachment; filename=\"graph_#{id}.zip\"")
+      |> send_resp(:ok, do_csv(attrs))
+    end
+  end
+
   defp do_drawing(%{"type" => "lineage", "ids" => [_ | _] = ids} = params) do
     Lineage.lineage(ids, options(params))
   end
@@ -45,10 +64,19 @@ defmodule TdDdWeb.GraphController do
 
   defp do_drawing(%{"type" => "sample"}), do: Lineage.sample()
 
+  defp do_csv(%{"type" => "lineage", "ids" => [_ | _] = ids} = params) do
+    Lineage.lineage_csv(ids, options(params))
+  end
+
+  defp do_csv(%{"type" => "impact", "ids" => [_ | _] = ids} = params) do
+    Lineage.impact_csv(ids, options(params))
+  end
+
   defp options(%{} = params) do
     []
     |> with_excludes(params)
     |> with_levels(params)
+    |> with_header_labels(params)
   end
 
   defp options(_params), do: []
@@ -58,4 +86,9 @@ defmodule TdDdWeb.GraphController do
 
   defp with_levels(acc, %{"levels" => levels}), do: acc ++ [levels: levels]
   defp with_levels(acc, _params), do: acc
+
+  defp with_header_labels(acc, %{"header_labels" => header_labels}),
+    do: acc ++ [header_labels: header_labels]
+
+  defp with_header_labels(acc, _params), do: acc
 end